added patch script
