+++ config/default.js
@@ -18,8 +18,6 @@ module.exports = {
   DYNAMODB: {
     endpoint: process.env.DYNAMODB_ENDPOINT || 'http://localhost:8000'
   },
-  API: {
-    TAGS_API_V3:process.env.TAGS_API_V3 || ""
-  },
+  TAGS_API_V3:process.env.TAGS_API_V3 || "",
   MAX_DAYS_FOR_CHALLENGE_SKILLS: parseInt(process.env.MAX_DAYS_FOR_CHALLENGE_SKILLS, 10) || 1
 }
