+++ package.json
@@ -21,6 +21,7 @@
     "luxon": "^1.26.0",
     "moment": "^2.29.1",
     "node-mb-string-size": "^1.0.4",
+    "prettier": "^2.2.1",
     "underscore": "^1.12.0",
     "winston": "^3.3.3"
   },
